+++ changelog.md
@@ -16,6 +16,7 @@ Check the command for more infos.
 * Added sun-resistant Drowned
   * Drowned have 30 seconds where they don't burn in sunlight. During this time, they will not seek water
 * Pillagers now try to reposition to not hit allies (or when they hit an ally)
+* Web throwing feature is no longer limited to spiders
 * Shulker bullets now take duration and amplifier from owners so each shulker can have its own
 * Snow golems can now strafe when attacking (disabled by default)
 * Fixed temporary cobwebs breaking any block in the cobweb's position